Routing Tweaks
Since Okapi is routing the requests to mod-kb-ebsco, addressing CORS in the Rack middleware doesn't actually make sense.  As such, we've yanked it.

The module descriptor was adjusted to accept all paths under the `/eholdings/` namespace.

The API Controller code was tweaked to slice the leading `/eholdings` from the path before passing through to RM-API.