Verify peer certificate when deploying module to okapi cluster
Deployments have been broken for this repository since we configured SSL on okapi.frontside.io.  To fix deployments we now keep the intermediate chain certificate in Travis CI as a base64 encoded environment variable.  During deployment we use this string to construct a .pem file that curl can use to verify the peer certificate.