Declare implementation of eholdings
The e-holdings application could potentially use multiple backends it
implement the management of digital resources. As such, it needs to
look up which service to use by interface name.

This makes the `mod-kb-ebsco module` provide the `eholdings` interface
so that it can be looked up and linked to the eholding UI app at runtime.