Use ActiveRecord-esque models for RMAPI Resources
This is a refactor of our previous JSONAPI work, as well as some new functionality.  The goal is to push all of the data loading from RMAPI into the model layer, thus substantially slimming down the controllers.  With this work, all RMAPI resources are now accessible via JSONAPI endpoints, and all relationships between the resources have been wired up.  Writes are not yet supported, but this is intended to be a feature-complete implementation of all the read operations needed by the frontend.