Scope configuration per-request.
There can be multiple requests happening simultaneously for different
tenants, and so we don't want to mutate the global class constants for
our flexirest resources. E.g. if `Title.configure` has side-effects,
then it could mess with another request in the same process.

This makes it so that each request gets its own subclass of a resource
configured just for that request, so no conflicts between tenants.