Apply ChunkedTransferEncoding Middleware to Rails 'development' environment
Jeffrey's original addition of the ChunkedTransferEncoding middleware does
successfully fix our issue with PUT requests to JSONAPI endpoints.  It was not
working before because apparently our Rails app is running in development mode
and the middelware had been configured for the production environment.  We
should configure Rails to run in production mode while on production, but for
now this will allow us to accept requests that respect strong parameters and
play nicely with the deserialization layer.