lazily fetch package in packages controller.
There is currently a `before_action` that eagerly fetches the package
on an explicit set of requests (show, update, destroy,
resources). This has a couple of disadvantages.

1. The list of requests has to be manually curated, and if at some
point, the package isn't needed for one of the requests, it could
still be doing the work of fetching it unecessarily.
2. It uses rails magic `before_action :set_package` where none is
really necessary.

Instead of magic, just use a regular ruby method that instantiates the
package lazily. This has the advantage of being normal Ruby, and also,
if only getting called when someone actually needs the package.