Link Resources to Custom Packages
Addresses UIEH-287.  This is backend work to support the linking of
both managed and custom titles to custom packages, thereby creating
custom resources.  To support this, the UI will need to be able to
filter packages by 'custom' anywhere package collections are dealt
with.  The UI will also need to be able to fire off a request that
will associate a custom package with a title and yield a unique,
compound ID representing the new "custom resource".

There's a pretty major refactor shoved in here too.  For all other
`filter` query parameters, RMAPI supports them and they are passed
through directly.  In this case we needed to pull large collections of
packages over the wire one page at a time and then handle the
filtering logic inside of mod-kb-ebsco. This is, of course,
inefficient, but unless you're hitting `/packages` directly (which
takes about a minute to return) it's pretty bearable if scoped to a
vendor or filtered even further with other query params (which are
supported in parallel).  The rather massive refactor afforded us the
ability to finely tune how data is persisted.  It also de-couples us
from several layers of the application that we've noticed slowing us
down for the past several months - namely Flexirest, whose AR inspired
ORM layer was increasingly strained by the complexity of RMAPI.  Here
we use a pattern somewhere between a data mapper and a domain-specific
repository.  This separates the storage of data from its manipulation,
leading to clean models (which now look more like Codex structures
than RMAPI structures), slimmer controllers, and an open door for
complex queries such as the one needed to populate the drop-down for
this feature.

For now this overhaul was kept to the package route, which is thus far
the only place we truly needed to break free Flexirest / ORM pattern.
There will be documentation provided showing the major changes
involved.