Add skeletal eholdings RAML.
As it stands, mod-kb-ebsco docs are not present in the api doc
repository here https://dev.folio.org/reference/api/ which is a shame.

This adds a skeletal RAML file so that we can deploy our documentation
to production. Then, after this is in place, we can add more
documentation for each endpoint.