Use fetch over require in strong params
In cases where a json payload it sent with empty attributes
using require will cause a failure within Ruby on
the ActionParams and bail out. What we rather want is for the
Strong Params to try to get the value and if not found set the
default value which will then continue on into the rest of the code
and hit the Validation block and return a appropriate error
and error code.