RAML Descriptors for Title Resources
We are using RAML to express our API contract in a commonly consumed
format that is spoken in one way or another by many modern web
development ensembles.  We can then publish this as a both a source of
human-readable documentation for the non `ui-eholdings` API consumer
and a schema that can be consumed by tooling to create an interface
betwen modules or even to draft an outline of a module - anything
available to RAML is now available as an option we can offer to users.

This work covers the endpoints related to Title resources, based at `/eholdings/titles`.