Fix type of the `providerToken` field
For some reason, the `type` of this field was an object containing an array: `{"type": ["object", "null"]}`. Unless there is some subtlety of JSON Schema that I don't know about, I think this must be a simple typo. At any rate, it prevented mod-graphql from properly ingesting the schema, so I have changed to to `"object"`, which seems from context to be correct.

(See https://issues.folio.org/browse/MODGQL-55 if interested in the history.)